Restructure TCP crates (#275)
The plan consists of restructuring TCP crates following best practices to simplify the understanding of each component.
