+++ huginn-net-tcp/src/mtu/extractor.rs
@@ -1,4 +1,4 @@
-use crate::observable::ObservableMtu;
+use super::observable::ObservableMtu;
 use pnet::packet::tcp::TcpFlags::SYN;
 use pnet::packet::tcp::TcpPacket;
 use tracing::debug;
