+++ huginn-net-tcp/src/mtu/observable.rs
@@ -1,0 +1,4 @@
+/// Observable MTU signals
+pub struct ObservableMtu {
+    pub value: u16,
+}
