+++ huginn-net-tcp/src/parser/hash.rs
