+++ huginn-net-tcp/src/tcp/mod.rs
@@ -1,3 +1,15 @@
+pub mod ip_options;
+pub mod observable;
+pub mod syn_options;
+pub mod ttl;
+pub mod window_size;
+
+pub use ip_options::IpOptions;
+pub use observable::{ObservableTcp, TcpObservation};
+pub use syn_options::{parse_options_raw, ParsedTcpOptions};
+pub use ttl::{calculate_ttl, guess_distance};
+pub use window_size::detect_win_multiplicator;
+
 use core::fmt;
 use std::fmt::Formatter;
