+++ huginn-net-tcp/src/tcp/syn_options.rs
@@ -4,7 +4,7 @@
 //! Pair it with [`crate::ttl::calculate_ttl`] and [`crate::window_size::detect_win_multiplicator`]
 //! to assemble a complete [`crate::observable::TcpObservation`].
 
-use crate::tcp::TcpOption;
+use super::TcpOption;
 use pnet::packet::tcp::{TcpOptionNumbers::*, TcpOptionPacket};
 use pnet::packet::{Packet, PacketSize};
