+++ huginn-net-tcp/src/uptime/observable.rs
@@ -1,0 +1,9 @@
+/// Observable Uptime signals
+#[derive(Debug, Clone)]
+pub struct ObservableUptime {
+    pub days: u32,
+    pub hours: u32,
+    pub min: u32,
+    pub up_mod_days: u32,
+    pub freq: f64,
+}
